Extract file_system module from system
