+++ yash-env/src/system/virtual/file_system.rs
@@ -244,23 +244,10 @@ impl FileBody {
     }
 }
 
-/// File permission bits.
-///
-/// The `Default` mode is `0o644`, not `0o000`.
-#[derive(Copy, Clone, Eq, Hash, PartialEq)]
-pub struct Mode(pub nix::sys::stat::mode_t);
-
-impl Debug for Mode {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "Mode({:#o})", self.0)
-    }
-}
-
-impl Default for Mode {
-    fn default() -> Mode {
-        Mode(0o644)
-    }
-}
+/// This type alias exists only for historical reasons.
+/// Please use `yash_env::system::Mode` instead.
+#[deprecated = "use yash_env::system::Mode instead"]
+pub use super::super::Mode2 as Mode;
 
 /// Implementor of [`Dir`] for virtual file system
 #[derive(Clone, Debug)]
